fix function sexp deserialization issue
I disregarded fun in sexpable record, serialization pretends to be fine,
but deserialization failed with raising an exception. Two functions
mattered: one is in State and the other is expire function in tcptimer.
The thing (on_close) in State is clear, but it's tricky for the tcptimer
currently used by delayed ack (ack.ml) and Tx segment (segment.ml). I
defer this issue until basic test is done, checking 'TODO-mig'.